mcp/servers/trading-journal/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trading-journal-mcp"
version = "0.1.0"
description = "MCP server: reads NeuroTrader SQLite, writes Obsidian markdown journal, optional Qdrant embeddings."
authors = [{ name = "Clayten Fields" }]
license = { text = "MIT" }
requires-python = ">=3.10"
dependencies = [
  "mcp>=1.0.0",
  "pyyaml>=6.0",
  "sqlalchemy>=2.0",
  "qdrant-client>=1.7",
  "httpx>=0.27",
]

[project.scripts]
trading-journal-mcp = "trading_journal.server:main"

[tool.hatch.build.targets.wheel]
packages = ["trading_journal"]
